requirements.txt
streamlit==1.56.0
numpy==1.26.4
tensorflow==2.15.0
pillow==10.3.0